ab_merkle_tree

Module balanced

Structs

BalancedMerkleTree

Merkle Tree variant that has hash-sized leaves and is fully balanced according to configured generic parameter.

Functions

compute_root_only_large_stack_size

Inner function used in BalancedMerkleTree::compute_root_only() for stack allocation, only public due to use in generic bounds

ensure_supported_n

Ensuring only supported N can be specified for BalancedMerkleTree.